requirements.txt
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psutil==6.0.0
pycparser==2.22
PyNaCl==1.5.0
PySocks==1.7.1
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.4
